Moved test constants out of `constants.js`
This just moves a couple of constant values which were only being used
for JS unit tests out of `constants.js` (and into `test_constants.js`).
Since we have non-test data in `constants.js` (meaning that file is
included in the bundled production app) we should get a little decrease
in bundle size.

closes #2348
pr #2354
