+++ static/js/components/dashboard/CourseRow_test.js
@@ -14,6 +14,8 @@ import {
   DASHBOARD_RESPONSE,
   COURSE_PRICES_RESPONSE,
   FINANCIAL_AID_PARTIAL_RESPONSE,
+} from '../../test_constants';
+import {
   STATUS_NOT_PASSED,
   STATUS_OFFERED,
   STATUS_MISSED_DEADLINE,
