+++ static/js/components/inputs/inputs_test.js
@@ -7,7 +7,7 @@ import VirtualizedSelect from 'react-virtualized-select';
 import R from 'ramda';
 import ga from 'react-ga';
 
-import { USER_PROFILE_RESPONSE } from '../../constants';
+import { USER_PROFILE_RESPONSE } from '../../test_constants';
 import iso3166 from 'iso-3166-2';
 import { modifyWrapperSelectField } from '../../util/test_utils';
 import SelectField from './SelectField';
