+++ static/js/containers/DashboardPage_test.js
@@ -33,6 +33,8 @@ import * as util from '../util/util';
 import {
   CYBERSOURCE_CHECKOUT_RESPONSE,
   EDX_CHECKOUT_RESPONSE,
+} from '../test_constants';
+import {
   TOAST_FAILURE,
   TOAST_SUCCESS,
